Add Rotary Encoder support
Based on:
https://www.circuitsathome.com/mcu/rotary-encoder-interrupt-service-routine-for-avr-micros

This assumes you are using a 2 bit quadrature rotary encoder. Right now
it is configured to only register detents (both A and B are N/C)
otherwise the counter will increase/decrease by 4 for each click. If
using a no-detent encoder than the return logic can be removed.